+++ src/constants/message.ts
@@ -6,6 +6,8 @@ const CART_MESSAGE = {
 const USER_MESSAGE = {
   FAIL_LOGIN: '아이디 혹은 비밀번호를 확인해주세요.',
   FAIL_SIGNUP: '기입하신 정보를 확인해주세요.',
+  FAIL_EDIT: '주소 혹은 핸드폰 번호를 확인해주세요.',
+  FAIL_WITH_DRAW: '들어올 땐 마음대로였지만 나갈 땐 아닐걸?',
   ASK_WITH_DRAW: '탈퇴하시겠습니까?',
 };
